prspec/requirements.txt
google-generativeai>=0.3.0
openai>=1.0.0
requests
tree-sitter
tree-sitter-python
tree-sitter-go
pyyaml
python-dotenv
click
rich
gitpython
